Added a font object for each GUI Widget. Added debug information to the GUI Widget, showing the hover and focus states. Added a method of showing the inheritance between GUI Widgets. Fixed a bug with incorrect mouse position modifications.

+++ BattleSysCommon/Graphics/GUI/GUIContainer.h
@@ -19,7 +19,7 @@ public:
 	GUIContainer(float x, float y, float width, float height);
 	virtual ~GUIContainer();
 	void add(GUIWidget* widget);
-	virtual void update(int mouseX, int mouseY);
+	virtual void update(double mouseX, double mouseY);
 	virtual void render();
 	virtual void deepDelete();
 };
